Big update. ACTHUNG: breaking EDIPY2... (admittedly I am not able to fix it).
* The code build has been revised:
two components: edipack2 and edipack2_cbindings, corresponding to a static and a dynamic library.

* The CMake now hosts a specific variable (see the list of variables with cmake .. -L)
WITH_INEQ=True/False, default=True
If True the inequivalent sites extension is included in the compilation.
The build leads to a single static library (as before) BUT two distinct modules
EDIPACK2
EDIPACK2INEQ
so one can safely work with the single-site module without having to worry and being sure that
wrong dimensions of some array WILL NOT be supported (i.e. will raise an error).

* Documentation has been slightly revised to this point to follow the new structure of the code
3 sections:
EDIPACK2: the library (as before plus adjustements due to the detach of RDMFT)
EDIPACK2INEQ: the inequivalent sites extension with a brief discussion
C-BINDING: a brief overview without further details, link to the source code for the interested reader.

TODO:
In EDIPACK2_C the C-Binding module (which has a specific pre-compilation flag) I included a new variable
has_ineq=0,1 c-bound to pass the information whether the dynamical library is or not compiled with ineq support.
This will make EDIPY2 works again. But it is failing and I do not know how to fix it.... shame on me.