Intermediate commit
Fixed a bug in the linker flags for libedipack2_cbinding.so. Using the
variable SCIFOR_LINK_LIBRARIES does not include absolute library paths
such as those of blas and lapack. Hence, if scalapack is not present
these are not linked. Using SCIFOR_LDFLAGS seems to fix the issue.
Tested on Debian (good day's work with @aamaricci)