Intermediate commit
Fixed a bug whereby the input list was not deleted upon finalization of
the solver. This would create an issue in cases where the solver is
initialized multiple times, and each time a differently named input file
is read. The linked list of input variables would continously grow.
Fixed by deleting the input in delete_ed_structure

Co-authored-by: alexkowalski <alexander.kowalski@uni-wuerzburg.de>