Fix in ED_MODE=SUPERC in case of MPI installation with NPH>0 and ED_SPARSE_H=T
The previous implementation was bugged, and indeed I had inserted time ago an ERROR STOP
if all the previous conditin were realized.
The actual implementation when NPH>0 does not take advantage of the tensor product of the Hamiltonian
between the phononic and electronic part but does the MPI slicing of the full DimEl*DimPh Fock space.
A smart implementation for the future could be slicing only the Electronic part but this would require a change
of order in the costruction of the states between the phononic and electronic one with the phonon index
being the one that cycle faster (opposite to now where it is the slowest one) so that one could
store the full Phononic Hamiltonians on each CPU and take advantage of the tensorization.

The actual implementation still works fine and when doing multiorbital cases one may want to avoid
storing the matrixes at the moment and use ED_SPARSE_H=F that has been the standard for me.